requirements.txt
pyzmq>=25.0
Pillow>=10.0
pygame>=2.5.0
numpy>=1.24
